Remove form settings from another file #JC19-455
