+++ apps/react/src/features/auth/components/RegistrationForm/RegistrationForm.tsx
@@ -1,34 +1,18 @@
-import * as yup from 'yup';
-import { memo, useEffect, useState } from 'react';
+import { memo, useCallback, useEffect, useState } from 'react';
 import { useFormik } from 'formik';
 import TextField from '@mui/material/TextField';
 import { Box, Grid, Snackbar, Typography } from '@mui/material';
 import { LoadingButton } from '@mui/lab';
+import { useNavigate } from 'react-router-dom';
 
 import { AppError } from '@js-camp/core/models/app-error';
 import { registrationUser } from '@js-camp/react/store/auth/dispatchers';
 import { useAppDispatch, useAppSelector } from '@js-camp/react/store/store';
-import { selectAreAuthLoading, selectError } from '@js-camp/react/store/auth/selectors';
+import { selectAreAuthLoading, selectError, selectIsSubmit } from '@js-camp/react/store/auth/selectors';
 
 import { ExtractedError, extractError } from '../../utils/error';
 
-interface FormData {
-
-  /** First name. */
-  readonly firstName: string;
-
-  /** Last name. */
-  readonly lastName: string;
-
-  /** Email. */
-  readonly email: string;
-
-  /** Password.  */
-  readonly password: string;
-
-  /** Password confirm. */
-  readonly passwordConfirm: string;
-}
+import { RegistrationFormData, signUpSchema } from './formSettings';
 
 const INITIAL_USER = {
   firstName: '',
@@ -42,47 +26,12 @@ const INITIAL_FORM_VALUE = {
   passwordConfirm: '',
 };
 
-const passwordPattern = /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d).{8,64}$/i;
-
-const FIRST_NAME_ERROR_MESSAGE = 'First name is required';
-const LAST_NAME_ERROR_MESSAGE = 'Last name is required';
-const EMAIL_ERROR_MESSAGES = {
-  email: 'Enter a valid email',
-  required: 'Email is required',
-};
-const PASSWORD_ERROR_MESSAGES = {
-  min: 'Password should be of minimum 8 characters length',
-  matches: 'Password must be at least 1 lowercase character, at least 1 uppercase character and at least 1 digit',
-  required: 'Password is required',
-};
-const PASSWORD_CONFIRM_ERROR_MESSAGES = {
-  oneOf: 'Passwords must match',
-  required: 'Password confirm is required',
-};
-
-const signUpSchema: yup.SchemaOf<FormData> = yup.object({
-  firstName: yup.string().required(FIRST_NAME_ERROR_MESSAGE),
-  lastName: yup.string().required(LAST_NAME_ERROR_MESSAGE),
-  email: yup
-    .string()
-    .email(EMAIL_ERROR_MESSAGES.email)
-    .required(EMAIL_ERROR_MESSAGES.required),
-  password: yup
-    .string()
-    .min(8, PASSWORD_ERROR_MESSAGES.min)
-    .matches(passwordPattern, PASSWORD_ERROR_MESSAGES.matches)
-    .required(PASSWORD_ERROR_MESSAGES.required),
-  passwordConfirm: yup
-    .string()
-    .oneOf([yup.ref('password'), null], PASSWORD_CONFIRM_ERROR_MESSAGES.oneOf)
-    .required(PASSWORD_CONFIRM_ERROR_MESSAGES.required),
-});
-
 export const RegistrationFormComponent = () => {
   const isLoading = useAppSelector(selectAreAuthLoading);
   const registrationError = useAppSelector(selectError);
+  const isSubmitForm = useAppSelector(selectIsSubmit);
   const dispatch = useAppDispatch();
-
+  const navigate = useNavigate();
   const [snackbar, setSnackbar] = useState({
     isOpen: false,
     message: '',
@@ -90,20 +39,16 @@ export const RegistrationFormComponent = () => {
   });
 
   useEffect(() => {
+    if (isSubmitForm) {
+      navigate('/');
+    }
+  }, [isSubmitForm]);
+
+  useEffect(() => {
     if (registrationError instanceof AppError) {
       setErrors(extractError(registrationError));
     }
   }, [registrationError]);
-
-  const handleSubmitForm = ({ email, firstName, lastName, password }: FormData) => {
-    dispatch(registrationUser({ email, firstName, lastName, password }));
-  };
-
-  const formik = useFormik({
-    initialValues: INITIAL_FORM_VALUE,
-    validationSchema: signUpSchema,
-    onSubmit: handleSubmitForm,
-  });
 
   const setErrors = (error: ExtractedError) => {
     formik.setErrors(
@@ -119,6 +64,16 @@ export const RegistrationFormComponent = () => {
     }
     setSnackbar(state => ({ ...state, isOpen: false }));
   };
+
+  const handleSubmitForm = useCallback(({ email, firstName, lastName, password }: RegistrationFormData) => {
+    dispatch(registrationUser({ email, firstName, lastName, password }));
+  }, [dispatch]);
+
+  const formik = useFormik({
+    initialValues: INITIAL_FORM_VALUE,
+    validationSchema: signUpSchema,
+    onSubmit: handleSubmitForm,
+  });
 
   return (
     <Box component="div">
